#72 refac: Views 폴더생성 및 MonthWidgetView 분리
